% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm.R
\name{predict_dfm}
\alias{predict_dfm}
\title{Predictions from an estimated dynamic factor model.}
\usage{
predict_dfm(data, output_dfm, months_ahead = 3, lag = 0)
}
\arguments{
\item{data}{matrix of variables, size (n_obs, n_variables). Must include in 1st column a series of type date, called "date", all data already stationary.}

\item{output_dfm}{list, the output of the \code{dfm()} function.}

\item{months_ahead}{number of months ahead to forecast.}

\item{lag}{number of lags for the kalman filter}
}
\value{
dataframe with all missing values filled + predictions.
}
\description{
runs a dataset through a previously estimated DFM to obtain predictions for all missing values in the series.
}
